Specialize disconnection events for ClientConnections
`DiconnectHandler` used to deal in `AbstractConnection` despite only ever being used with `ClientConnection`
objects. Additionally, only `ClientConnection` can fire disconnection events, so the methods for registering
`DisconnectHandler` has been moved from `Connection` to `ClientConnection`. `Connection.fireDisconnect()` is actually an
internal utility method and is no longer exposed publicly.
